fixed deleteUser()

+++ project-group-13-Backend/src/main/java/ca/mcgill/ecse321/projectgroup13/services/UserService.java
@@ -145,7 +145,8 @@ public class UserService {
         	if (order.getOrderStatus().equals(OrderStatus.PaymentPending))
         		orderService.deleteOrder(order);
         }
-        cartRepository.delete(cart);
+        if (user.getCart() != null)
+            cartRepository.delete(cart);
         userRepository.delete(user);
     }
